New Appointments linked with Messages and MessageDisplay Update
Also fixed a call in the views.py of the message app to correctly retrieve the full name of users. Updated json package to fix vulnerabilities and added a package (dompurify). I did not receive any errors with this update, but check your version of this project to be sure when you pull these changes.